requirements.txt
pymongo>=4.6.1
python-dotenv>=1.0.0 
yfinance>=0.2.14
pandas>=2.1.1
numpy>=1.26.1
matplotlib>=3.8.0
requests>=2.31.0
beautifulsoup4>=4.12.2
scikit-learn>=1.3.0
scipy>=1.11.3
langchain-community>=0.2.1
openai
selenium
Pillow>=9.0.0
playwright>=1.31.0
torch>=2.0.0
transformers>=4.28.0
ragatouille>=0.0.1
pytz>=2022.7
sentencepiece




